# Conflicts:
#	src/main/java/com/alrex/parcool/proxy/ClientProxy.java

+++ src/main/java/com/alrex/parcool/client/animation/impl/FastSwimAnimator.java
@@ -26,7 +26,6 @@ public class FastSwimAnimator extends Animator {
         float tick = getTick() + transformer.getPartialTick();
         double armFactor = Math.cos(Math.PI * tick / 15.);
 
-
         HumanoidArm attackHand = player.getMainArm();
         boolean leftArmAnimatable = attackHand != HumanoidArm.LEFT || transformer.getRawModel().attackTime <= 0f;
         boolean rightArmAnimatable = attackHand != HumanoidArm.RIGHT || transformer.getRawModel().attackTime <= 0f;
@@ -54,15 +53,15 @@ public class FastSwimAnimator extends Animator {
         if (rightArmXAngle < 0) rightArmXAngle = -3 * Math.sqrt(-rightArmXAngle);
         if (leftArmXAngle < 0) leftArmXAngle = -3 * Math.sqrt(-leftArmXAngle);
         if (rightArmAnimatable) {
-            transformer.getRawModel().rightArm.x += Math.max(0, 2.4 * Math.sin(Math.PI * tick / 15.));
-            transformer.getRawModel().rightArm.y += Math.max(0, 1.2 * Math.sin(Math.PI * tick / 15.));
-            transformer.getRawModel().rightArm.z -= 1.2 * Math.sin(Math.PI * tick / 15.);
+            transformer.getRawModel().rightArm.x += (float) Math.max(0, 2.4 * Math.sin(Math.PI * tick / 15.));
+            transformer.getRawModel().rightArm.y += (float) Math.max(0, 1.2 * Math.sin(Math.PI * tick / 15.));
+            transformer.getRawModel().rightArm.z -= (float) (1.2 * Math.sin(Math.PI * tick / 15.));
         }
 
         if (leftArmAnimatable) {
-            transformer.getRawModel().leftArm.x -= Math.max(0, -2.4 * Math.sin(Math.PI * tick / 15.));
-            transformer.getRawModel().leftArm.y += Math.max(0, -1.2 * Math.sin(Math.PI * tick / 15.));
-            transformer.getRawModel().leftArm.z += 1.2 * Math.sin(Math.PI * tick / 15.);
+            transformer.getRawModel().leftArm.x -= (float) Math.max(0, -2.4 * Math.sin(Math.PI * tick / 15.));
+            transformer.getRawModel().leftArm.y += (float) Math.max(0, -1.2 * Math.sin(Math.PI * tick / 15.));
+            transformer.getRawModel().leftArm.z += (float) (1.2 * Math.sin(Math.PI * tick / 15.));
         }
 
         transformer.getRawModel().head.z += 0.5f * animationFactor;
